requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.2.0
Pillow>=10.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
